move dtype device mixin to more general place
